+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   When a job is selected in the overview tab, show the workers for that job in the work panel.
 -   Show progress bars on livestock tab.
 -   Show an alert when an AI manager has been constructed and there are still manager's desks constructed.
+-   Made it so that you can have overrides of default values per animal type for the Livestock jobs.
 
 ### Changed